Separate CouchConnection from CouchInterpreter.
