+++ src/test/scala/com/ironcorelabs/davenport/DBSpec.scala
@@ -6,12 +6,9 @@
 package com.ironcorelabs.davenport
 
 import scalaz._, Scalaz._, scalaz.concurrent.Task
-import org.scalatest.{ WordSpec, Matchers, BeforeAndAfterAll }
-import org.typelevel.scalatest._
-import scala.language.postfixOps
 import DB._
 
-class DBSpec extends WordSpec with Matchers with BeforeAndAfterAll with DisjunctionMatchers {
+class DBSpec extends TestBase {
   "DB" should {
     "fail lifting none into dbprog" in {
       MemInterpreter.interpretTask(liftIntoDBProg(None)).run._2 should be(left)
